Remove unused files and general clean-up (#11)
* Remove unused files and code
* Add missing `__init__.py`s
* Rename `algos` -> `algorithms` and `envs` -> `environments`
* Overall clean-up here and there
