+++ .gitignore
@@ -1,5 +1,3 @@
-*__pycache__*
-.idea
-rllab/config_personal.py
 data
-vendor/mujoco
+*.pyc
+.idea
