+++ softqlearning/environments/__init__.py
@@ -1,0 +1,1 @@
+from .multigoal import MultiGoalEnv
