--- softqlearning/envs/mujoco/swimmer_env.py
@@ -1,75 +1,0 @@
-from rllab.envs.base import Step
-from rllab.misc.overrides import overrides
-from rllab.envs.mujoco.mujoco_env import MujocoEnv
-import numpy as np
-from rllab.core.serializable import Serializable
-from rllab.misc import logger
-from rllab.misc import autoargs
-
-
-class SwimmerEnv(MujocoEnv, Serializable):
-
-    FILE = 'swimmer.xml'
-    ORI_IND = 2
-
-    @autoargs.arg('ctrl_cost_coeff', type=float,
-                  help='cost coefficient for controls')
-    def __init__(
-            self,
-            ctrl_cost_coeff=1e-2,
-            undirected=False,
-            *args, **kwargs):
-        self.ctrl_cost_coeff = ctrl_cost_coeff
-        self.undirected = undirected
-        self.progress_threshold = 1
-        super(SwimmerEnv, self).__init__(*args, **kwargs)
-        Serializable.quick_init(self, locals())
-
-    def get_current_obs(self):
-        return np.concatenate([
-            self.model.data.qpos.flat,
-            self.model.data.qvel.flat,
-            self.get_body_com("torso").flat,
-        ]).reshape(-1)
-
-    def get_ori(self):
-        return self.model.data.qpos[self.__class__.ORI_IND]
-
-    def step(self, action):
-        self.forward_dynamics(action)
-        next_obs = self.get_current_obs()
-        lb, ub = self.action_bounds
-        scaling = (ub - lb) * 0.5
-        ctrl_cost = 0.5 * self.ctrl_cost_coeff * np.sum(
-            np.square(action / scaling))
-        forward_reward = self.get_body_comvel("torso")[0]
-        if self.undirected:
-            forward_reward = abs(forward_reward)
-        reward = forward_reward - ctrl_cost
-        done = False
-        return Step(next_obs, reward, done)
-
-    @overrides
-    def log_diagnostics(self, paths):
-        if len(paths) > 0:
-            progs = [
-                path["observations"][-1][-3] - path["observations"][0][-3]
-                for path in paths
-            ]
-            logger.record_tabular('AverageForwardProgress', np.mean(progs))
-            logger.record_tabular('MaxForwardProgress', np.max(progs))
-            logger.record_tabular('MinForwardProgress', np.min(progs))
-            logger.record_tabular('StdForwardProgress', np.std(progs))
-
-            if self.undirected:
-                n_directions = [
-                    np.max(progs) > self.progress_threshold,
-                    np.min(progs) < - self.progress_threshold,
-                    ].count(True)
-                logger.record_tabular('nDirections', n_directions)
-        else:
-            logger.record_tabular('AverageForwardProgress', np.nan)
-            logger.record_tabular('MaxForwardProgress', np.nan)
-            logger.record_tabular('MinForwardProgress', np.nan)
-            logger.record_tabular('StdForwardProgress', np.nan)
-
